initial comment; Added components and context folder for clean-up;
